Resolve #1 - Add the 'smart model binder' to the CSF.Web project
This is an ASP.NET MVC2 model binder based on the work at:
http://lostechies.com/jimmybogard/2009/03/18/a-better-model-binder/

There are still some issues with this, specifically a unit test that is
currently marked as ignored.  This can be improved at some point so that
multiple instances of the same model binder type are counted as duplicates
instead of being silently accepted.